+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 * Added support for automatically advancing the dialog.
+* Added `GetVars` to `RumorScope` as a way for getting all of the variables.
 
 ### Removed
 * `Exodrifter.Rumor.Engine.Util` is now internal instead of public.